requirements.txt
gunicorn
psycopg2-binary
asgiref==3.3.4
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
iniconfig==1.1.1
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.2.0
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4